podvm: Enable se image build for rhel with and without --no-verify option
Made changes to build RHEL SE enabled podvm image with an extra Arg of SE_VERIFY=true

+++ src/cloud-api-adaptor/podvm/qcow2/build-s390x-se-image.sh
@@ -7,15 +7,42 @@ elif [ "${ARCH}" != "s390x" ]; then
     exit 0
 fi
 echo "Building SE podvm image for $ARCH"
-echo "Finding host key files"
-host_keys=""
-rm /tmp/files/.dummy.crt || true
-for i in /tmp/files/*.crt; do
-    [[ -f "$i" ]] || break
-    echo "found host key file: \"${i}\""
-    host_keys+="-k ${i} "
+
+SE_VERIFY=${SE_VERIFY:-true}
+
+readonly BASE_DIR="/tmp/files"
+if [ "${SE_VERIFY}" = "true" ]; then
+	required_files=("DigiCertCA.crt" "ibm-z-host-key-gen2.crl" "ibm-z-host-key-signing-gen2.crt")
+	for file in "${required_files[@]}"; do
+        local_path="${BASE_DIR}/${file}"
+        if [[ -f "${local_path}" ]]; then
+            echo "Found required file: ${local_path}"
+        else
+            echo "Missing required file: ${local_path}" >&2
+            exit 1
+   		fi
+    done
+    signcert="${BASE_DIR}/ibm-z-host-key-signing-gen2.crt"
+    cacert="${BASE_DIR}/DigiCertCA.crt"
+    crl="${BASE_DIR}/ibm-z-host-key-gen2.crl"
+fi
+
+declare -a host_key_args=()
+shopt -s nullglob
+found_hkd=0
+for file in "${BASE_DIR}"/*HKD.crt; do
+    if [[ -f "$file" ]]; then
+        echo "Found HKD file: \"${file}\""
+        host_key_args+=("-k" "${file}")
+        found_hkd=1
+    fi
 done
-[[ -z $host_keys ]] && echo "Didn't find host key files, please download host key files to 'files' folder " && exit 1
+shopt -u nullglob
+if [[ "$found_hkd" -eq 0 ]]; then
+    echo "No HKD files found in ${BASE_DIR}." >&2
+    exit 1
+fi
+rm /tmp/files/.dummy.crt || true
 
 if [ "${PODVM_DISTRO}" = "rhel" ]; then
     export LANG=C.UTF-8
@@ -190,12 +217,19 @@ echo "Generating an IBM Secure Execution image"
 echo "Creating SE boot image"
 export SE_PARMLINE="root=/dev/mapper/$LUKS_NAME rd.auto=1 rd.retry=30 console=ttysclp0 quiet panic=0 rd.shell=0 blacklist=virtio_rng swiotlb=262144"
 sudo -E bash -c 'echo "${SE_PARMLINE}" > ${dst_mnt}/boot/parmfile'
+
+if [ "${SE_VERIFY}" = "true" ]; then
+    EXTRA_ARGS=" --cert=${cacert} --cert=${signcert} --crl=${crl} "
+else
+    EXTRA_ARGS=" --no-verify "
+fi
+
 sudo -E /usr/bin/genprotimg \
     -i ${dst_mnt}/boot/${KERNEL_FILE} \
     -r ${dst_mnt}/boot/${INITRD_FILE} \
     -p ${dst_mnt}/boot/parmfile \
-    --no-verify \
-    ${host_keys} \
+    "${host_key_args[@]}" \
+    ${EXTRA_ARGS} \
     -o ${dst_mnt}/boot-se/se.img
 
 # exit and throw an error if no se image was created
